requirements.txt
cryptography>=41.0.0
pycryptodome>=3.19.0
qiskit>=0.45.0
qiskit-aer>=0.13.0
liboqs-python>=0.9.0
numpy>=1.24.0
